string_id moves
- rename string_ids.h to string_id.h
- rename HaloStrings.h to global_string_ids.h and move to tag_files folder
- rename e_global_string_ids to e_global_string_id
