+++ xlive/Blam/Cache/DataTypes/BlamDataTypes.h
@@ -5,7 +5,7 @@
 // simply just include the headers you need in the blam data type header
 // and use this header for external usage
 
-#include "Blam/Engine/tag_files/string_ids.h"
+#include "Blam/Engine/tag_files/string_id.h"
 #include "Blam/Engine/cseries/cseries_strings.h"
 #include "TagRef.h"
 #include "BlamTag.h"
