+++ xlive/Blam/Engine/game/cheats.cpp
@@ -1,7 +1,7 @@
 #include "stdafx.h"
 #include "cheats.h"
 
-#include "Blam/Enums/HaloStrings.h"
+#include "Blam/Engine/tag_files/global_string_ids.h"
 #include "Blam/Engine/game/game_globals.h"
 
 #include "H2MOD/Tags/TagInterface.h"
